requirements.txt
psycopg2-binary
pydantic
sqlalchemy
uvicorn

fastapi == 0.81.0
